Move Dynamic and Static into http.resource
Resources are an HTTP concept, so it makes sense to put them under
aspen.http. Then aspen.resources is really about loading and caching,
though I'm not sure that doesn't eventually want to get pushed down into
simplates itself (decode_raw, e.g., certainly belongs in simplates).
